chore: Connect or create OrderReceipt in OrdersService
The code changes in `orders.service.ts` modify the `OrdersService` class to connect or create an `OrderReceipt` based on the `paidOrderDto`. If an `OrderReceipt` with the corresponding `orderId` exists, it will be connected. Otherwise, a new `OrderReceipt` will be created with the provided `receiptUrl`. This improves the handling of `OrderReceipt` in the service.

Note: The recent user commits and repository commits are not relevant for generating the commit message.